Feature: Trace stack during object initialization
-- This is needed because classes like Serial are used bym any
high-level classes like MagneticStirrer and Centrifuge, each of which
may correspondto a separate module on the deck. Now, the initialization
arguments do contain separate network addresses, like COM16 and COM8 for
COM ports, which can help us distinguish between these. But then we do
not have a mapping of network addresses to deck modules. Therefore, to
automate the process, I am also tracing the entire stack at the time of
initialization. Thus, the stack trace for Serial may show that it was
initialized fir a MagneticStirrer or a Centrigure and so on.